+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Breaking changes:
 
 - Remove :owner and :repo, and rename :git_http_url and :git_http_userinfo [#1411](https://github.com/sider/runners/pull/1411)
 - Replace :pull_number with :refspec [#1412](https://github.com/sider/runners/pull/1412)
+- Unify `S3_ENDPOINT` to `RUNNERS_OPTIONS` [#1414](https://github.com/sider/runners/pull/1414)
 
 ## 0.33.0